routes gefixed

+++ dwengo_backend/controllers/teacher/teacherAssignmentController.ts
@@ -3,10 +3,20 @@ import { AuthenticatedRequest } from "../../interfaces/extendedTypeInterfaces";
 import teacherAssignmentService from "../../services/teacherServices/teacherAssignmentService";
 
 export class AssignmentTeacherController {
+
+    private isUserValid(req: AuthenticatedRequest, res: Response): boolean {
+        if (!req.user) {
+            res.status(401).json({ error: "Unauthorized: No user found" });
+            return false;
+        }
+        return true;
+    }
+
     async createAssignmentForClass(req: AuthenticatedRequest, res: Response): Promise<void> {
         try {
+            if (!this.isUserValid(req, res)) return;
             const { classId, learningPathId, deadline }: { classId: number, learningPathId: number, deadline: string } = req.body;
-            const teacherId: number = Number(req.user.id);
+            const teacherId: number = Number(req.user!.id);
 
             const parsedDeadline = new Date(deadline);
 
@@ -19,8 +29,9 @@ export class AssignmentTeacherController {
 
     async getAssignmentsByClass(req: AuthenticatedRequest, res: Response): Promise<void> {
         try {
+            if (!this.isUserValid(req, res)) return;
             const classId: number = parseInt(req.params.classId);
-            const teacherId: number = Number(req.user.id);
+            const teacherId: number = Number(req.user!.id);
             const assignments = await teacherAssignmentService.getAssignmentsByClass(classId, teacherId);
             res.status(200).json(assignments);
         } catch (error) {
@@ -30,9 +41,10 @@ export class AssignmentTeacherController {
 
     async updateAssignment(req: AuthenticatedRequest, res: Response): Promise<void> {
         try {
+            if (!this.isUserValid(req, res)) return;
             const assignmentId: number = parseInt(req.params.assignmentId);
             const learningPathId: number = req.body;
-            const teacherId: number = Number(req.user.id);
+            const teacherId: number = Number(req.user!.id);
             const updatedAssignment = await teacherAssignmentService.updateAssignment(assignmentId, learningPathId, teacherId);
             res.json(updatedAssignment);
         } catch (error) {
@@ -42,8 +54,9 @@ export class AssignmentTeacherController {
 
     async deleteAssignment(req: AuthenticatedRequest, res: Response): Promise<void> {
         try {
+            if (!this.isUserValid(req, res)) return;
             const assignmentId: number = parseInt(req.params.assignmentId);
-            const teacherId: number = Number(req.user.id);
+            const teacherId: number = Number(req.user!.id);
             await teacherAssignmentService.deleteAssignment(assignmentId, teacherId);
             res.status(204).send();
         } catch (error) {
